改进上传过程中的文件验证机制

+++ kc/kc_donjin/lib.py
@@ -3,11 +3,25 @@ import rarfile
 from kc_donjin.config import *
 
 
+class UploadedFileExists(Exception):
+    pass
+
+
+class UploadedFileFormatError(Exception):
+    pass
+
+
+class UploadedFileContentError(Exception):
+    pass
+
+
 def handle_uploaded_file(f):
     fn = f.name
     if fn.find('.') == -1:
         fn += '.rar'
     path = os.path.join(KC_DONJIN_UPLOAD_DIR, fn)
+    if os.path.exists(path):
+        raise UploadedFileExists
 
     with open(path, 'wb+') as destination:
         for chunk in f.chunks():
@@ -22,5 +36,9 @@ def handle_uploaded_file(f):
                     cnt += 1
         if cnt >= 2:
             return path
+        else:
+            os.unlink(path)
+            raise UploadedFileContentError
+
     os.unlink(path)
-    return False
+    raise UploadedFileFormatError
